% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{backend_is_installed}
\alias{backend_is_installed}
\title{Check if Torch Backend Libraries are Installed}
\usage{
backend_is_installed(...)
}
\arguments{
\item{...}{Optional parameters passed to \code{\link[torch:torch_is_installed]{torch::torch_is_installed()}}.}
}
\value{
Boolean indicating whether the LibTorch and LibLantern backend libraries are installed.
}
\description{
Check if Torch Backend Libraries are Installed
}
\examples{
if(backend_is_installed()){
  cat("torch's LibTorch and LibLantern backend libraries are installed!\n")
}
}
