% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQRN.R
\name{batch_size_default}
\alias{batch_size_default}
\title{Default batch size (internal)}
\usage{
batch_size_default(tensor_dat, batch_size = 256)
}
\arguments{
\item{tensor_dat}{A \code{\link[torch:dataset]{torch::dataset()}}.}

\item{batch_size}{An initial batch size, by default \code{256}.}
}
\value{
The fixed batch_size.
}
\description{
Default batch size (internal)
}
\examples{
#TODO
}
