% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{check_backend_installed}
\alias{check_backend_installed}
\title{Check if Torch Backends are Installed}
\usage{
check_backend_installed(behaviour = c("error", "bool"))
}
\arguments{
\item{behaviour}{one of "error" or "bool".}
}
\value{
Boolean indicating whether the LibTorch and LibLantern backends are installed.
If \code{behaviour = "error"}, an error is raised if the backends are not installed.
}
\description{
Check if Torch Backends are Installed
}
\keyword{internal}
