% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_X_matrix}
\alias{check_X_matrix}
\title{Check the simulation X matrix}
\usage{
check_X_matrix(X, n, p)
}
\arguments{
\item{X}{Covariate matrix.}

\item{n}{Number of observations.}

\item{p}{Number of covariates.}
}
\value{
Returns TRUE if X is a matrix with dimension n * p. Otherwise an error is raised.
}
\description{
Check the simulation X matrix
}
\keyword{internal}
