% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQRN.R
\name{decay_learning_rate}
\alias{decay_learning_rate}
\title{Performs a learning rate decay step on an optimizer}
\usage{
decay_learning_rate(optimizer, decay_rate)
}
\arguments{
\item{optimizer}{A \code{torch::optimizer} object.}

\item{decay_rate}{Learning rate decay factor.}
}
\value{
The \code{optimizer} with a decayed learning rate.
}
\description{
Performs a learning rate decay step on an optimizer
}
\examples{
#TODO
}
