% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQRN.R
\name{default_device}
\alias{default_device}
\title{Default torch device}
\usage{
default_device()
}
\value{
Returns \code{torch::torch_device("cuda")} if \code{torch::cuda_is_available()}, or \code{torch::torch_device("cpu")} otherwise.
}
\description{
Default torch device
}
\examples{
if(backend_is_installed()){
  device <- default_device()
  print(device)
}
}
