% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{ensure_backend_installed}
\alias{ensure_backend_installed}
\title{Ensure Torch Backend Libraries are Installed}
\usage{
ensure_backend_installed(
  behaviour = c("error", "warn", "message", "bool"),
  ...
)
}
\arguments{
\item{behaviour}{one of "error" or "bool".}

\item{...}{Optional parameters passed to \code{\link[torch:torch_is_installed]{torch::torch_is_installed()}}.}
}
\value{
Boolean indicating whether the LibTorch and LibLantern backend libraries are installed.
If \code{behaviour = "error"}, an error is raised if the backend libraries are not installed.
}
\description{
Ensure Torch Backend Libraries are Installed
}
\keyword{internal}
