% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQRN-package.R
\docType{package}
\name{EQRN-package}
\alias{EQRN}
\alias{EQRN-package}
\title{EQRN: Extreme Quantile Regression Neural Networks}
\description{
A user-friendly framework for forecasting and extrapolating extreme measures of conditional risk using flexible neural network architectures. It allows for capturing complex multivariate dependencies, including dependencies between observations, such as sequential dependence (time-series).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/opasche/EQRN}
  \item \url{https://opasche.github.io/EQRN/}
  \item Report bugs at \url{https://github.com/opasche/EQRN/issues}
}

}
\keyword{internal}
