% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQRN.R
\name{EQRN_load}
\alias{EQRN_load}
\title{Load an EQRN object from disc}
\usage{
EQRN_load(path, name = NULL, device = default_device(), ...)
}
\arguments{
\item{path}{Path to the save location as a string.}

\item{name}{String name of the save.
If \code{NULL} (default), assumes the save name has been given implicitly in the \code{path}.}

\item{device}{(optional) A \code{\link[torch:torch_device]{torch::torch_device()}}. Defaults to \code{\link[=default_device]{default_device()}}.}

\item{...}{DEPRECATED. Used for back-compatibility.}
}
\value{
The loaded \code{"EQRN"} model.
}
\description{
Loads in memory an \code{"EQRN"} object that has previously been saved on disc using \code{\link[=EQRN_save]{EQRN_save()}}.
}
\examples{
#TODO
}
