% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQRN.R
\name{EQRN_save}
\alias{EQRN_save}
\title{Save an EQRN object on disc}
\usage{
EQRN_save(fit_eqrn, path, name = NULL, no_warning = TRUE)
}
\arguments{
\item{fit_eqrn}{An \code{"EQRN"} object}

\item{path}{Path to save folder as a string.}

\item{name}{String name of the save.}

\item{no_warning}{Whether to silence the warning raised if a save folder needed beeing created (bool).}
}
\value{
Nothing
}
\description{
Creates a folder named \code{name} and located in \code{path}, containing binary save files,
so that the given \code{"EQRN"} object \code{fit_eqrn} can be loaded back in memory from disc using \code{\link[=EQRN_load]{EQRN_load()}}.
}
