% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{excess_probability}
\alias{excess_probability}
\title{Excess Probability Predictions}
\usage{
excess_probability(object, ...)
}
\arguments{
\item{object}{A model object for which excess probability prediction is desired.}

\item{...}{additional model-specific arguments affecting the predictions produced.
See the corresponding method documentation.}
}
\description{
A generic function (method) for excess probability predictions from various fitted EQR models.
The function invokes particular methods which depend on the class of the first argument.
}
