% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQRN.R
\name{fix_dimsimplif}
\alias{fix_dimsimplif}
\title{(INTERNAL) Corrects a dimension simplification bug from the torch package}
\usage{
fix_dimsimplif(dl_i, ..., responses = TRUE)
}
\arguments{
\item{dl_i}{batch object from an itteration over a \code{\link[torch:dataloader]{torch::dataloader()}}.}

\item{...}{dimension(s) of the covariate object (excluding the first "batch" dimension)}

\item{responses}{Bolean indicating whether the batch object \code{dl_i} is a covariates-response pair.}
}
\value{
The fixed dl_i object
}
\description{
(INTERNAL) Issue was raised to the \code{torch} maintainers and should be fixed, deprecating this function.
}
\examples{
#TODO
}
\keyword{internal}
