% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_backend}
\alias{install_backend}
\title{Install Torch Backend}
\usage{
install_backend(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[torch:install_torch]{torch::install_torch()}}.}
}
\description{
This function can be called just after installing the EQRN package.
It installs the necessary LibTorch backend by calling \code{\link[torch:install_torch]{torch::install_torch()}}.
See \url{https://torch.mlverse.org/docs/articles/installation.html} for more details.
Calling this function shouldn't be necessary in most cases, as loading EQRN
(e.g. with \code{library(EQRN)} or \code{EQRN::fct()}) should do it automatically (via \code{.onLoad()}).
}
\examples{
\dontrun{
EQRN::install_backend()
}
}
