% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{last_elem}
\alias{last_elem}
\title{Last element of a vector}
\usage{
last_elem(x)
}
\arguments{
\item{x}{Vector.}
}
\value{
The last element in the vector \code{x}.
}
\description{
Returns the last element of the given vector in the most efficient way.
}
\details{
The last element is obtained using \code{x[length(x)]}, which is done in \code{O(1)} and faster than, for example, any of
\code{Rcpp::mylast(x)}, \code{tail(x, n=1)}, \code{dplyr::last(x)}, \verb{x[end(x)[1]]]}, and \code{rev(x)[1]}.
}
\examples{
last_elem(c(2, 6, 1, 4))
}
