% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQRN.R
\name{legacy_names}
\alias{legacy_names}
\title{Internal renaming function for back-compatibility}
\usage{
legacy_names(eqrn_fit, classes = NULL)
}
\arguments{
\item{eqrn_fit}{EQRN fitted object.}

\item{classes}{If provided, overrides classes of \code{eqrn_fit}.}
}
\value{
The \code{eqrn_fit} object with updated attribute names and classes.
}
\description{
Internal renaming function for back-compatibility
}
\keyword{internal}
