% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list2matrix}
\alias{list2matrix}
\title{Convert a list to a matrix}
\usage{
list2matrix(lst, dim = c("row", "col"))
}
\arguments{
\item{lst}{A list.}

\item{dim}{One of \code{"row"} (default) or \code{"col"}.}
}
\value{
The list converted to a matrix, by stacking the elements of \code{lst} in the rows or columns of a matrix.
}
\description{
Convert a list to a matrix
}
\examples{
#list2matrix(list(2, 7, 3, 8), "row")
}
