% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_folds}
\alias{make_folds}
\title{Create cross-validation folds}
\usage{
make_folds(y, num_folds, stratified = FALSE)
}
\arguments{
\item{y}{Numerical vector of observations}

\item{num_folds}{Number of folds to create.}

\item{stratified}{Logical value. If \code{TRUE}, the folds are stratified along \code{rank(y)}.}
}
\value{
Vector of indices of the assigned folds for each observation.
}
\description{
Utility function to create folds of data, used in cross-validation proceidures.
The implementation is from the \code{gbex} \code{R} package
}
\examples{
make_folds(rnorm(30), 5)
}
