% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrix2list}
\alias{matrix2list}
\title{Convert a matrix to a list}
\usage{
matrix2list(mat)
}
\arguments{
\item{mat}{A matrix.}
}
\value{
A list with elements corresponding to rows of \code{mat}.
}
\description{
Convert a matrix to a list
}
\examples{
#matrix2list(matrix(c(1, 2, 3, 4, 5, 6), ncol=2))
}
