% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_metrics.R
\name{multilevel_prop_below}
\alias{multilevel_prop_below}
\title{Multilevel 'proportion_below'}
\usage{
multilevel_prop_below(
  y,
  Pred_Q,
  proba_levels,
  prefix = "",
  na.rm = FALSE,
  give_names = TRUE
)
}
\arguments{
\item{y}{Vector of observations.}

\item{Pred_Q}{Matrix of of size \code{length(y)} times \code{proba_levels},
whose columns are the quantile predictions at each \code{proba_levels} and
each row corresponds to an observation or realisation.}

\item{proba_levels}{Vector of probability levels at which the predictions were made.
Must be of length \code{ncol(Pred_Q)}.}

\item{prefix}{A string prefix to add to the output's names (if \code{give_names} is \code{TRUE}).}

\item{na.rm}{A logical value indicating whether \code{NA} values should be stripped before the computation proceeds.}

\item{give_names}{Whether to name the output proportions (bool).}
}
\value{
A vector of length \code{length(proba_levels)} giving the proportion of observations
below the predictions (\code{Pred_Q}) at each probability level.
If \code{give_names} is \code{TRUE}, the output vector is named \code{paste0(prefix, "propBelow_q", proba_levels)}.
}
\description{
Multilevel version of \code{\link[=proportion_below]{proportion_below()}}.
}
