% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQRN_network_structures.R
\name{nn_alpha_dropout}
\alias{nn_alpha_dropout}
\title{Alpha-dropout module}
\usage{
nn_alpha_dropout(p = 0.5, inplace = FALSE)
}
\arguments{
\item{p}{probability for dropout.}

\item{inplace}{whether the dropout in performed inplace.}
}
\description{
An alpha-dropout layer as a \code{\link[torch:nn_module]{torch::nn_module}}, used in self-normalizing networks.
}
\details{
The constructor allows specifying:
\itemize{
\item{p}{probability of an element to be zeroed (default is 0.5),}
\item{inplace}{if set to TRUE, will do the operation in-place (default is FALSE).}
}
}
\references{
Gunter Klambauer, Thomas Unterthiner, Andreas Mayr, Sepp Hochreiter. Self-Normalizing Neural Networks.
Advances in Neural Information Processing Systems 30 (NIPS 2017), 2017.
}
