% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQRN_network_structures.R
\name{nn_dropout_nd}
\alias{nn_dropout_nd}
\title{Dropout module}
\usage{
nn_dropout_nd(p = 0.5, inplace = FALSE)
}
\arguments{
\item{p}{probability for dropout.}

\item{inplace}{whether the dropout in performed inplace.}
}
\description{
A dropout layer as a \code{\link[torch:nn_module]{torch::nn_module}}.
}
\details{
The constructor allows specifying:
\itemize{
\item{p}{probability of an element to be zeroed (default is 0.5),}
\item{inplace}{if set to TRUE, will do the operation in-place (default is FALSE).}
}
}
