% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_metrics.R
\name{proportion_below}
\alias{proportion_below}
\title{Proportion of observations below conditional quantile vector}
\usage{
proportion_below(y, Q_hat, na.rm = FALSE)
}
\arguments{
\item{y}{Vector of observations.}

\item{Q_hat}{Vector of predicted quantiles.}

\item{na.rm}{A logical value indicating whether \code{NA} values should be stripped before the computation proceeds.}
}
\value{
The proportion of observation below the predictions.
}
\description{
Proportion of observations below conditional quantile vector
}
\examples{
proportion_below(c(2.3, 4.2, 1.8), c(2.9, 5.6, 1.7))
}
