% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_metrics.R
\name{quantile_prediction_error}
\alias{quantile_prediction_error}
\title{Quantile prediction calibration error}
\usage{
quantile_prediction_error(y, Q_hat, prob_level, na.rm = FALSE)
}
\arguments{
\item{y}{Vector of observations.}

\item{Q_hat}{Vector of predicted quantiles at probability level \code{prob_level}.}

\item{prob_level}{Probability level of the predicted quantile.}

\item{na.rm}{A logical value indicating whether \code{NA} values should be stripped before the computation proceeds.}
}
\value{
The quantile prediction error calibration metric.
}
\description{
Quantile prediction calibration error
}
\examples{
quantile_prediction_error(c(2.3, 4.2, 1.8), c(2.9, 5.6, 2.7), prob_level=0.8)
}
