% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_metrics.R
\name{R_squared}
\alias{R_squared}
\title{R squared}
\usage{
R_squared(y, y_hat, na.rm = FALSE)
}
\arguments{
\item{y}{Vector of observations or ground-truths.}

\item{y_hat}{Vector of predictions.}

\item{na.rm}{A logical value indicating whether \code{NA} values should be stripped before the computation proceeds.}
}
\value{
The R squared of the predictions \code{y_hat} for \code{y}.
}
\description{
The coefficient of determination, often called R squared, is the proportion of data variance explained by the predictions.
}
\examples{
R_squared(c(2.3, 4.2, 1.8), c(2.2, 4.6, 1.7))
}
