% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rep_tibble}
\alias{rep_tibble}
\title{Tibble replicatior}
\usage{
rep_tibble(tbl, m)
}
\arguments{
\item{tbl}{A \code{\link[tibble:tibble]{tibble::tibble()}}.}

\item{m}{An integer.}
}
\value{
The tibble is replicated \code{m} times and colums names appended with \code{rep_id = 1:m}.
}
\description{
Tibble replicatior
}
\keyword{internal}
