% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rep_vector2matrix}
\alias{rep_vector2matrix}
\title{Replicated vector to matrix}
\usage{
rep_vector2matrix(vec, nrep, dim = c("row", "col"))
}
\arguments{
\item{vec}{Vector.}

\item{nrep}{Number of repetitions.}

\item{dim}{One of \code{"row"} (default) or \code{"col"}.}
}
\value{
Matrix of replicated vector.
}
\description{
Replicated vector to matrix
}
\examples{
#rep_vector2matrix(c(2, 7, 3, 8), 3, "row")
}
