% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roundm}
\alias{roundm}
\title{Mathematical number rounding}
\usage{
roundm(x, decimals = 0)
}
\arguments{
\item{x}{Vector of numerical values to round.}

\item{decimals}{Integer indicating the number of decimal places to be used.}
}
\value{
A vector containing the entries of \code{x}, rounded to \code{decimals} decimals.
}
\description{
This function rounds numbers in the mathematical sense,
as opposed to the base \code{R} function \code{\link[=round]{round()}} that rounds 'to the even digit'.
}
\examples{
roundm(2.25, 1)
}
