% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQRN_seq.R
\name{setup_optimizer_seq}
\alias{setup_optimizer_seq}
\title{Instantiate an optimizer for training an EQRN_seq network}
\usage{
setup_optimizer_seq(network, learning_rate, L2_pen, optim_met = "adam")
}
\arguments{
\item{network}{A \code{torch::nn_module} network to be trained in \code{\link[=EQRN_fit_seq]{EQRN_fit_seq()}}.}

\item{learning_rate}{Initial learning rate for the optimizer during training of the neural network.}

\item{L2_pen}{L2 weight penalty parameter for regularization during training.}

\item{optim_met}{DEPRECATED. Optimization algorithm to use during training. \code{"adam"} is the default.}
}
\value{
A \code{torch::optimizer} object used in \code{\link[=EQRN_fit_seq]{EQRN_fit_seq()}} for training.
}
\description{
Instantiate an optimizer for training an EQRN_seq network
}
\examples{
#TODO
}
\keyword{internal}
