% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_metrics.R
\name{square_loss}
\alias{square_loss}
\title{Square loss}
\usage{
square_loss(y, y_hat)
}
\arguments{
\item{y}{Vector of observations or ground-truths.}

\item{y_hat}{Vector of predictions.}
}
\value{
The vector of square errors between \code{y} and \code{y_hat}.
}
\description{
Square loss
}
\examples{
square_loss(c(2.3, 4.2, 1.8), c(2.2, 4.6, 1.7))
}
