% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{start_doParallel_strategy}
\alias{start_doParallel_strategy}
\title{Start a doParallel execution strategy}
\usage{
start_doParallel_strategy(
  strategy = c("sequential", "parallel"),
  n_workers = NULL
)
}
\arguments{
\item{strategy}{One of \code{"sequential"} (default) or \code{"parallel"}.}

\item{n_workers}{Number of parallel workers as an integer.
Defaults to \code{\link[parallel:detectCores]{parallel::detectCores()}}\code{-1} if \code{NULL} (default).
Ignored if \code{strategy=="sequential"}.}
}
\value{
A named list containing:
\itemize{
\item{par_operator}{the relevant \code{\link[foreach:foreach]{foreach::foreach()}} loop operator,}
\item{cl}{the cluster object.}
}
}
\description{
Start a doParallel execution strategy
}
\examples{
\dontrun{
cl <- start_doParallel_strategy("parallel", n_workers=3)
stop_doParallel_strategy("parallel", cl)
}
}
\keyword{internal}
