% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop_doParallel_strategy}
\alias{stop_doParallel_strategy}
\title{Stop the current doParallel strategy}
\usage{
stop_doParallel_strategy(strategy = c("sequential", "parallel"), cl)
}
\arguments{
\item{strategy}{One of \code{"sequential"} (default) or \code{"parallel"}.}

\item{cl}{Cluster object, returned by \code{\link[=start_doParallel_strategy]{start_doParallel_strategy()}}, called with the same \code{strategy}.}
}
\description{
Stops the given cluster, using \code{\link[parallel:makeCluster]{parallel::stopCluster()}}, if \code{strategy=="parallel"}.
}
\keyword{internal}
