% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vec2mat}
\alias{vec2mat}
\title{Convert a vector to a matrix}
\usage{
vec2mat(v, axis = c("col", "row"))
}
\arguments{
\item{v}{Vector.}

\item{axis}{One of \code{"col"} (default) or \code{"row"}.}
}
\value{
The vector \code{v} as a matrix.
If \code{axis=="col"} (default) the column vector \code{v} is returned as a \code{length(v)} times \code{1} matrix.
If \code{axis=="row"}, the vector \code{v} is returned as a transposed \code{1} times \code{length(v)} matrix.
}
\description{
Convert a vector to a matrix
}
\examples{
vec2mat(c(2, 7, 3, 8), "col")
}
