% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{onload_backend_installer}
\alias{onload_backend_installer}
\title{On-Load Torch Backend Internal Install helper}
\usage{
onload_backend_installer(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[torch:install_torch]{torch::install_torch()}}.}
}
\description{
On-Load Torch Backend Internal Install helper
}
\keyword{internal}
